Simplify the verification mechanism.
